ComfyUI-LHM/requirements.txt
torch>=2.3.0
torchvision>=0.18.0
numpy>=1.23.0
Pillow>=11.1.0
opencv-python
rembg>=2.0.63
smplx
basicsr==1.4.2
kornia==0.7.2
timm==1.0.15
transformers>=4.41.2
accelerate
omegaconf>=2.3.0
pyrender>=0.1.45
trimesh>=4.4.9 